Add files via upload
In this version, I added a feature that removes the text and image when you are not in proximity to any of the locations, and instead displays the derryfield logo. I also added text to speech for the text, enabling the text to be read aloud with the click of a button. I also made a small change by rounding off the feet values so the GUI doesn't look as cluttered